Tab bar fixes (#860)
* :lipstick: Add `TaBarPresenter`

* 📐Fix tabBar colors
