+++ teferi/UI/Modules/Intro/IntroPresenter.swift
@@ -32,32 +32,8 @@ class IntroPresenter : NSObject
     
     func showMainScreen()
     {
-        let imageInsets = UIEdgeInsetsMake(6, 0, -6, 0)
-        let tabBarController = UITabBarController()
-        tabBarController.tabBar.backgroundColor = .white
-        tabBarController.tabBar.barTintColor = .white
-        
-        let topline = CALayer()
-        topline.frame = CGRect(x: 0, y: 0, width: tabBarController.tabBar.frame.width, height: 0.5)
-        topline.backgroundColor = UIColor(r: 240, g: 240, b: 240).cgColor
-        tabBarController.tabBar.layer.addSublayer(topline)
-        
+        let tabBarController = TabBarPresenter.create(with: viewModelLocator)
         tabBarController.transitioningDelegate = self
-        
-        let navigationController = NavigationPresenter.create(with: viewModelLocator)
-        navigationController.tabBarItem.image = Asset.home.image
-        navigationController.tabBarItem.imageInsets = imageInsets
-        
-        let goalNavigationController = GoalNavigationPresenter.create(with: viewModelLocator)
-        goalNavigationController.tabBarItem.image = Asset.goals.image
-        goalNavigationController.tabBarItem.imageInsets = imageInsets
-        
-        let weeklySummary = WeeklySummaryPresenter.create(with: viewModelLocator)
-        weeklySummary.tabBarItem.image = Asset.icChart.image
-        weeklySummary.tabBarItem.imageInsets = imageInsets
-        
-        tabBarController.viewControllers = [goalNavigationController, navigationController, weeklySummary]
-
         viewController.present(tabBarController, animated: true)
     }
 }
